Refactor src directory structure
